+++ CHANGELOG.md
@@ -100,7 +100,8 @@ All notable changes to this project will be documented in this file.
 - Added Terminus Machina for Party and Leader Option for Experiment Purposes
 - Made the Terminate Function more function faster than before
 - Added Check Notifications
-- Added Holia Keldrasil for dungeon automation
+- Added Holia Keldrasil for dungeon automation (Development Phase)
+- Added Altar of Siena B1F - Prideus for Development Phase
 
 ## v5.11 Fixes